fix(build): satisfy IParameterizable on net471 and widen mock ParameterCount
Two failures, both outside the net10.0 src build that local verification covers.

net471 has no default interface implementations, so IParameterizable declares
SupportsParameterInitialization and SanitizeParameters abstract on that target
while every other framework inherits them. Three types relied on the default and
so compiled everywhere except net471: DecisionTreeRegressionBase and
AsyncDecisionTreeRegressionBase were missing SupportsParameterInitialization,
and OnnxFullModelAdapter was missing both. Each now states them explicitly. The
ONNX adapter reports false: the graph arrives already trained and exposes no
settable parameters through this surface.

IParameterSource.ParameterCount is long, but four mock models in the benchmark
and testconsole projects still declared it int, which is CS0738 rather than a
silent narrowing. Widened to long.

Neither project is built by a net10.0 src-only verification, which is why both
survived to CI.

+++ AiDotNetBenchmarkTests/Helpers/MockNeuralNetwork.cs
@@ -44,7 +44,7 @@ public class MockNeuralNetwork<T, TInput, TOutput> : IFullModel<T, TInput, TOutp
         _parameters = parameters.Clone();
     }
 
-    public int ParameterCount => _parameters.Length;
+    public long ParameterCount => _parameters.Length;
     public bool SupportsParameterInitialization => ParameterCount > 0;
 
     public IFullModel<T, TInput, TOutput> WithParameters(Vector<T> parameters)
